requirements.txt
acdh-tei-pyutils
jinja2
pandas
tqdm
acdh-transkribus-utils